vm: Microsoft Visual Studio port work in progress

+++ vm/factor.cpp
@@ -120,7 +120,7 @@ void factor_vm::init_factor(vm_parameters *p)
 	if(p->image_path == NULL)
 		p->image_path = default_image_path();
 
-	srand(system_micros());
+	srand((unsigned int)system_micros());
 	init_ffi();
 	init_stacks(p->ds_size,p->rs_size);
 	init_callbacks(p->callback_size);
